# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZZIK LIVE API
  description: Location-based rewards and verification platform API
  version: 1.0.0
  contact:
    name: ZZIK Team
    email: api@zzik.live
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.zzik.live
    description: Production server
...
paths:
...
  /api/wallet/redeem:
    post:
      summary: Redeem a voucher (idempotent)
      tags: [Wallet]
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
          description: Unique key for idempotent operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - voucherId
              properties:
                voucherId:
                  type: string
      responses:
        '200':
          description: Voucher redeemed successfully
          headers:
            X-Idempotent-Replay:
              schema:
                type: string
                enum: ['0', '1']
              description: Whether this was a replay of previous request
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Rate limit maximum
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in window
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when limit resets
            Server-Timing:
              schema:
                type: string
              description: Performance timing
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  state:
                    type: string
                    enum: [used]
                  usedAt:
                    type: string
                    format: date-time
                  voucherId:
                    type: string
                  offerId:
                    type: string
        '422':
          description: Invalid voucher or parameters
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retry
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...